DBF修改與怪物資料由DBF提供

+++ Client/Assets/Script/Define/DBFFeature.cs
@@ -7,8 +7,8 @@ public class DBFFeature : DBF
 	public int StrID = 0; // 字串編號
 	public int Mode = 0; // 模式
 	public int Group = 0; // 群組
-	public int Chances = 0; // 機率
-	public int Cooldown = 0; // 冷卻時間
+	public float Chances = 0; // 機率
+	public float Cooldown = 0; // 冷卻時間
 	public int Effect = 0; // 效果
 	public int Value = 0; // 效果值
 }
